pyproject.toml
[project]
name = "vkgptbot"
version = "1.0.0"
description = "NedoGPT for Telegram and VK"
readme = "README.md"
requires-python = ">=3.12.10"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.16.4",
    "jinja2>=3.1.6",
    "loguru>=0.7.3",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "telegrinder>=0.5.1",
    "tiktoken>=0.9.0",
    "unidecode>=1.4.0",
    "vkbottle>=4.5.2",
]
